requirements.txt
pyautogui
torch
pytesseract
opencv
pillow